llm_client.py
# llm_client.py
import os
import cohere
from dotenv import load_dotenv

load_dotenv()

COHERE_API_KEY = os.getenv("COHERE_API_KEY")

if not COHERE_API_KEY:
    raise RuntimeError("COHERE_API_KEY not found in environment variables")

co = cohere.Client(COHERE_API_KEY)
