requirements.txt
fastapi
uvicorn
pydantic
chromadb
cohere
python-dotenv